fix: remove unused google-auto-auth@0.8.2
Removes unused package google-auto-auth@0.8.2. This package was listed as a direct dependency, but is not used in the package directly. This package is a dependency of `@google-cloud/storage` but since updating that package, we no longer need this version.

Removing this package removes 6 vulnerable paths to base64url@2.0.0 which is a High Severity vulnerability.
https://snyk.io/vuln/npm:base64url:20180511