Refactor server to run in all (Node.js compatible) runtimes and meta frameworks (#710)
Refactor the request handlers and hooks to be based on `Request` (from the fetch API) instead of `http.IncomingMessage`/`http.ServerResponse` (which are Node.js specific) as we can turn the Node.js ones into `Request`/`Response` but not the other way around. That means we can have a `handle(req, res)` method for Node.js servers and a `handleWeb(req)` for other runtimes and meta-frameworks (such as Next.js React Router, SvelteKit, etc) which both use the same core logic.